Added the model operator and changed the name of the 'saving_name'
