+++ plato/utils/checkpoint_operator.py
@@ -1,5 +1,5 @@
 """
-Save and load each personalized checkpoints in persoanlized federated learning.
+Save and load desired checkpoints.
 """
 import os
 import re
@@ -7,7 +7,7 @@ from typing import Dict, Optional, List
 
 import torch
 
-from plato.utils.format_saving_name import get_format_name
+from plato.utils.filename_formatter import get_format_name
 
 
 class CheckpointsOperator:
@@ -220,22 +220,3 @@ def perform_client_checkpoint_loading(
 
     return filename, cpk_oper
 
-
-def reset_all_weights(model: torch.nn.Module) -> None:
-    """
-    refs:
-    - https://discuss.pytorch.org/t/how-to-re-set-alll-parameters-in-a-network/20819/6
-    - https://stackoverflow.com/questions/63627997/reset-parameters-of-a-neural-network-in-pytorch
-    - https://pytorch.org/docs/stable/generated/torch.nn.Module.html
-    """
-
-    @torch.no_grad()
-    def weight_reset(module: torch.nn.Module):
-        # - check if the current module has reset_parameters & if it's callabed called it on m
-        reset_parameters = getattr(module, "reset_parameters", None)
-        if callable(reset_parameters):
-            module.reset_parameters()
-
-    # Applies fn recursively to every submodule see:
-    # https://pytorch.org/docs/stable/generated/torch.nn.Module.html
-    model.apply(fn=weight_reset)
